requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
uvloop==0.20.0
openai==1.45.0
